Moved tests from CreatAccTest and SignInTest to new class (NavigationTest). More suitable classification and better organisation.
